Only make the Nginx config if the app has domains
On top of being sensible (otherwise you have an nginx.conf that
specifies an upstream and does nothing with it), this also makes it
cleaner to have a Plushu configuration with apps that don't get
vhosted (say, because they're a non-web service that is bound directly
to the host interface). Just only set explicit domains for the apps
you want vhosted (either removing the app-name-domain plugin or the
APP_DOMAIN file to prevent non-vhosted apps from getting an implicit
domain).